tests/wpt/web-platform-tests/url/historical.html

<!doctype html>
<meta charset=utf-8>
<title>URL historical tests</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
var objects = [
  [function() { return window.location }, "location object"],
  [function() { return document.createElement("a") }, "a element"],
  [function() { return document.createElement("area") }, "area element"],
];

objects.forEach(function(o) {
  test(function() {
    var object = o[0]();
    assert_false("searchParams" in object,
                 o[1] + " should not have a searchParams attribute");
  }, "searchParams on " + o[1]);
});

test(function() {
  var url = new URL("./foo", "http://www.example.org");
  assert_equals(url.href, "http://www.example.org/foo");
  assert_throws(new TypeError(), function() {
    url.href = "./bar";
  });
}, "Setting URL's href attribute and base URLs");
</script>
